ai_backend/requirements.txt
# Core ML and Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Deep Learning (optional for more advanced models)
tensorflow==2.15.0
keras==2.15.0

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# CORS for frontend integration
fastapi-cors==0.0.6

# Model serialization
joblib==1.3.2
pickle-mixin==1.0.2

# Data validation
pydantic==2.5.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
